services/identity-service/pyproject.toml
[project]
name = "identity-service"
version = "0.1.0"
description = "Identity & Authorization Source of Truth"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlmodel>=0.0.22",
    "psycopg>=3.2.0",
    "alembic>=1.13.0",
    "boto3>=1.34.0",
    "pydantic-settings>=2.0.0",
    "pyjwt>=2.10.0",
    "cryptography>=43.0.0",
    "email-validator>=2.2.0",
    "structlog>=24.0.0",
]

[dependency-groups]
test = [
    "pytest>=8.3.0",
    "httpx>=0.27.0",
    "testcontainers[postgres]>=4.8.0",
]

[tool.uv]
managed = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = [".", ".."]
addopts = "-v --tb=short -p no:warnings"